requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2-binary
pydantic
python-multipart
passlib[bcrypt]
python-jose[cryptography]
alembic
email-validator
aiofiles
python-dotenv
fastapi-mail
celery
redis
boto3
strawberry-graphql[fastapi]